+++ components/src/lib/hooks/useKeyDown.ts
@@ -1,0 +1,17 @@
+import { useEffect } from "react";
+
+export default function useKeyDown(key: string, callback: () => void) {
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === key) {
+        callback();
+      }
+    };
+
+    window.addEventListener("keydown", handleKeyDown);
+
+    return () => {
+      window.removeEventListener("keydown", handleKeyDown);
+    };
+  }, []);
+}
